requirements.txt
Flask==2.3.3
Flask-SQLAlchemy==3.0.5
Flask-SocketIO==5.3.6
python-dotenv==1.0.0
eventlet==0.33.3
cryptography==41.0.4
python-multipart==0.0.6
gunicorn==21.2.0
psycopg2-binary==2.9.7
setuptools>=60.0.0